Given a (CDK-synth'd, YAML) CFn template, run local SFn & Lambda.
This is a simple script to set up a local Step Functions and Lambda test
environment for a CDK-based project.  It assumed that the local Step Functions
and Lambda systems so graciously provided by AWS in the form the Step Functions
local Docker image (amazon/aws-stepfunctions-local), and the SAM CLI (for Lambda
local), have already been set up.

The State Machine definition is extracted from the `synth'd` CDK template,
localised and injected into the local environment.  If everythings comes up
successfully, it can simply be run using the ARN printed when it is created:

  aws --region eu-central-1 stepfunctions \
      --endpoint http://localhost:8083 \
      start-execution \
      --state-machine arn:aws:states:us-east-1:123456789012:stateMachine:BuildReleaseStateMachine165B927F \
      --input "$(cat event.json)"

See:
 * https://docs.aws.amazon.com/step-functions/latest/dg/sfn-local.html
 * https://docs.aws.amazon.com/step-functions/latest/dg/sfn-local-config-options.html
 * https://docs.aws.amazon.com/serverless-application-model/latest/developerguide/sam-cli-command-reference-sam-local-start-lambda.html
 * https://docs.docker.com/docker-for-mac/networking/#use-cases-and-workarounds